Fixes responses minor bugs and misc refactoring
